feature(metrics) Move exchangesV2 group to _onchain_labels category
